requirements.txt
attrs==22.1.0
backoff==2.1.2
black==22.10.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
conductor-python==1.0.46 
coverage==6.5.0
cryptography==38.0.4
exceptiongroup==1.0.0
fastapi==0.85.1
filelock==3.8.2
flake8==5.0.4
frozenlist==1.3.1
h11==0.14.0
idna==3.4
iniconfig==1.1.1
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
multidict==6.0.2
mypy-extensions==0.4.3
oscrypto==1.3.0
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
pycodestyle==2.9.1
pycparser==2.21
pycryptodomex==3.16.0
pydantic==1.10.2
pyflakes==2.5.0
PyJWT==2.6.0 
pyOpenSSL==22.1.0 
pyparsing==3.0.9
pytest==7.2.0
pytest-cov==4.0.0
pytz==2022.6
python-dotenv==0.21.0
redmail==0.4.1
requests==2.28.1
requests-toolbelt==0.9.1
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.19.0
yarl==1.8.1
